--- src/stores.js
@@ -1,2 +1,0 @@
-import { writable } from "svelte/store";
-export const status = writable('dummy');
